+++ crates/rome_service/src/configuration/parse/json/rules.rs
@@ -1662,6 +1662,7 @@ impl VisitNode<JsonLanguage> for Nursery {
                 "noRedundantRoles",
                 "noSelfAssign",
                 "noStaticOnlyClass",
+                "noUselessEmptyExport",
                 "noVoid",
                 "useAriaPropTypes",
                 "useArrowFunction",
@@ -2100,6 +2101,29 @@ impl VisitNode<JsonLanguage> for Nursery {
                         diagnostics,
                     )?;
                     self.no_static_only_class = Some(rule_configuration);
+                }
+                _ => {
+                    diagnostics.push(DeserializationDiagnostic::new_incorrect_type(
+                        "object or string",
+                        value.range(),
+                    ));
+                }
+            },
+            "noUselessEmptyExport" => match value {
+                AnyJsonValue::JsonStringValue(_) => {
+                    let mut configuration = RuleConfiguration::default();
+                    self.map_to_known_string(&value, name_text, &mut configuration, diagnostics)?;
+                    self.no_useless_empty_export = Some(configuration);
+                }
+                AnyJsonValue::JsonObjectValue(_) => {
+                    let mut rule_configuration = RuleConfiguration::default();
+                    rule_configuration.map_rule_configuration(
+                        &value,
+                        name_text,
+                        "noUselessEmptyExport",
+                        diagnostics,
+                    )?;
+                    self.no_useless_empty_export = Some(rule_configuration);
                 }
                 _ => {
                     diagnostics.push(DeserializationDiagnostic::new_incorrect_type(